+++ src/services/projects/application/service.ts
@@ -1,57 +1,45 @@
 import { Injectable } from '@nestjs/common';
-import { FindOrder } from 'libs/orm';
 import { CreateBodyDto, ListQueryDto } from '../dto';
 import { ProjectRepository } from '../infrastructure/repository';
-import { Project, SortType } from '../domain/model';
+import { Project } from '../domain/model';
 import { Transactional } from '../../../libs/orm/transactional';
 import { User } from '../../users/domain/model';
 import { Role } from '../../roles/domain/model';
 import { RoleRepository } from '../../roles/infrastructure/repository';
-
-function getSort(sort?: SortType): FindOrder {
-  switch (sort) {
-    case 'Hot':
-      return { order: { bookMark: 'DESC', id: 'DESC' } };
-    case 'Oldest':
-      return { order: { id: 'ASC' } };
-    case 'Latest':
-    default:
-      return { order: { id: 'DESC' } };
-  }
-}
 
 @Injectable()
 export class ProjectService {
   constructor(private projectRepository: ProjectRepository, private roleRepository: RoleRepository) {}
 
   async list(args: ListQueryDto) {
-    const projects = await this.projectRepository.find(
-      {
-        stacks: args.stacks,
-        purpose: args.purpose,
-        job: args.job,
-        status: args.status,
-      },
-      {
-        limit: Number(args.limit),
-        page: Number(args.page),
-      },
-      getSort(args.sort),
-    );
-
-    const count = await this.projectRepository.count(
-      {
-        stacks: args.stacks,
-        purpose: args.purpose,
-        job: args.job,
-        status: args.status,
-      },
-      {
-        limit: Number(args.limit),
-        page: Number(args.page),
-      },
-      getSort(args.sort),
-    );
+    const [projects, count] = await Promise.all([
+      this.projectRepository.find(
+        {
+          stacks: args.stacks,
+          purpose: args.purpose,
+          job: args.job,
+          status: args.status,
+        },
+        {
+          limit: Number(args.limit),
+          page: Number(args.page),
+        },
+        args.sort,
+      ),
+      this.projectRepository.count(
+        {
+          stacks: args.stacks,
+          purpose: args.purpose,
+          job: args.job,
+          status: args.status,
+        },
+        {
+          limit: Number(args.limit),
+          page: Number(args.page),
+        },
+        args.sort,
+      ),
+    ]);
 
     return { projects, count };
   }
