+++ src/services/projects/infrastructure/repository.ts
@@ -3,8 +3,20 @@ import { Raw } from 'typeorm';
 import { JobType } from '../../roles/domain/model';
 import { FindOrder, PaginationOption, convertOptions, In } from '../../../libs/orm';
 import { Repository } from '../../../libs/ddd';
-import { Project, PurposeType, StatusType } from '../domain/model';
+import { Project, PurposeType, SortType, StatusType } from '../domain/model';
 import { stripUndefined } from '../../../libs/common';
+
+function getSort(sort?: SortType): FindOrder {
+  switch (sort) {
+    case 'Hot':
+      return { order: { bookMarkCount: 'DESC', id: 'DESC' } };
+    case 'Oldest':
+      return { order: { id: 'ASC' } };
+    case 'Latest':
+    default:
+      return { order: { id: 'DESC' } };
+  }
+}
 
 @Injectable()
 export class ProjectRepository extends Repository<Project, Project['id']> {
@@ -13,7 +25,7 @@ export class ProjectRepository extends Repository<Project, Project['id']> {
   async find(
     conditions: { stacks?: string[]; purpose?: PurposeType; job?: JobType; status?: StatusType },
     options?: PaginationOption,
-    order?: FindOrder,
+    sort?: SortType,
   ): Promise<Project[]> {
     return this.getManager().find(Project, {
       where: {
@@ -30,14 +42,14 @@ export class ProjectRepository extends Repository<Project, Project['id']> {
         }),
       },
       ...convertOptions(options),
-      ...order,
+      ...getSort(sort),
     });
   }
 
   async count(
     conditions: { stacks?: string[]; purpose?: PurposeType; job?: JobType; status?: StatusType },
     options?: PaginationOption,
-    order?: FindOrder,
+    sort?: SortType,
   ): Promise<number> {
     return this.getManager().count(Project, {
       where: {
@@ -54,7 +66,7 @@ export class ProjectRepository extends Repository<Project, Project['id']> {
         }),
       },
       ...convertOptions(options),
-      ...order,
+      ...getSort(sort),
     });
   }
 }
